changed searchBar to a component and added a push history to the property gallery page
