+++ src/components/LandingPage/LandingPage.jsx
@@ -1,26 +1,17 @@
 import React, { useState } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import './LandingPage.css';
-import CheckInCheckOut from '../SearchBar/CheckInCheckOut';
 // ----------------MUI-------------------
 import { Card, Grid } from '@mui/material';
 import { Button } from '@mui/material'
-import SearchIcon from '@mui/icons-material/Search';
-import Travelers from '../SearchBar/Travelers';
-import WhereTo from '../SearchBar/WhereTo';
+import SearchBar from '../SearchBar/SearchBar';
 
 
 function LandingPage() {
   const [heading, setHeading] = useState('A Traveled Path Homes');
   const history = useHistory();
-  const [datesPopup, setDatesPopup] = useState(false);
-  const [travelersPopup, setTravelersPopup] = useState(false);
-  const [wherePopup, setWherePopup] = useState(false);
-  const [locationPopup, setLocationPopup] = useState(false);
 
-
-  const bookingReducer = useSelector((store) => store.bookingReducer);
 
   const onLogin = (event) => {
     history.push('/login');
@@ -34,44 +25,12 @@ function LandingPage() {
     history.push('/propertygallery');
   };
 
+  
   // ---------------------DIALOG-MUI-----------------------------------
 
   return (
     <div className="containers">
-      <Card sx={{
-        padding: 2,
-        paddingLeft: 2,
-        paddingRight: 2,
-        fontSize: 30,
-        marginLeft: 40,
-        marginRight: 50,
-          height: 38,
-          width: 900
-      }}>
-        <Grid 
-        display='flex'
-        justifyContent='space-around'
-        >
-          <Button sx={{color: '#121957'}} 
-          onClick={() => setLocationPopup(true)}className="btn btn_sizeSm" >
-                  {bookingReducer.hospital}
-          </Button>
-          <Button sx={{color: '#121957'}} onClick={() => setDatesPopup(true)} className="btn btn_sizeSm" >
-                  Check In - Check Out
-          </Button>
-          <Button sx={{color: '#121957'}} onClick={() => setTravelersPopup(true)} className="btn btn_sizeSm" >
-                  How many Travelers?
-          </Button>
-          <Button sx={{color: '#121957'}} onClick={() => setButtonPopup(true)} className="btn btn_sizeSm" >
-                  Filters
-          </Button>
-          <SearchIcon></SearchIcon>
-        </Grid>
-      </Card>
-      <CheckInCheckOut trigger={datesPopup} setTrigger={setDatesPopup}/>
-      <Travelers trigger={travelersPopup} setTrigger={setTravelersPopup}/>
-      <WhereTo trigger={locationPopup} setTrigger={setLocationPopup}/>
-
+     <SearchBar />
       <br></br>
       <Card sx={{
         padding: 4,
